A Fish Feed Calculator program
The contents of fish feeds are calculated based on the age of the fish group being considered. Fish at early stages of their life are generally fed feed with higher protein content compared to in the later stages of their lives. This feed calculator will calculate the amount in kilogram of each component of a typical feed for any of 3 stages of a fish life; at juven ile, growout and broodstock. It will demand two inputs; age of fish group in weeks, and total kilogram of feed to be produced.